refactor: move service package to resources package
