+++ internal/entrypoints/mcpserver/tools/test_helpers_test.go
@@ -8,7 +8,7 @@ import (
 	"github.com/swarm-deploy/swarm-deploy/internal/gitops/differ/diff"
 	gitx "github.com/swarm-deploy/swarm-deploy/internal/gitops/git"
 	"github.com/swarm-deploy/swarm-deploy/internal/registry"
-	"github.com/swarm-deploy/swarm-deploy/internal/service"
+	"github.com/swarm-deploy/swarm-deploy/internal/resources/service"
 	"github.com/swarm-deploy/swarm-deploy/internal/swarm"
 )
