+++ internal/resources/service/stype/dict.go
